[FE] feat : 채팅 알림 추가 (미완성)
- socket을 통해 alertList에 추가되고, 페이지에 나타남
- x를 누르면 v-show:false가 됨
<보충할 점>
- 페이지를 벗어나면 alertList가 추기화되서 휘발됨
- 채팅을 눌렀을 때 채팅방으로 이동해야 함

S08P12E105-271

+++ FE/tonny-bunny/src/store/actions/alert.js
@@ -68,6 +68,10 @@ export default {
         console.log("알림 읽음 확인 수정 API", userSeq, alertLogSeq);
         console.log(payload);
 
+        if (alertLogSeq == -1 || alertLogSeq == "-1") {
+            console.log("Skip putAlert(chatting)");
+            return;
+        }
         let { data } = await http.put(`/alerts/log/${alertLogSeq}`);
 
         try {
@@ -101,6 +105,10 @@ export default {
         console.log("삭제 API", userSeq, alertLogSeq);
         console.log(payload);
 
+        if (alertLogSeq == -1 || alertLogSeq == "-1") {
+            console.log("Skip deleteAlert(chatting)");
+            return;
+        }
         let { data } = await http.delete(`/alerts/log/${alertLogSeq}`);
 
         try {
